[refactor-http-server] stream factored out
